chore(x/sophon): move transfer logic into a separate file
Part-of: #595
