+++ x/sophon/keeper/msg_server.go
@@ -1,7 +1,6 @@
 package keeper
 
 import (
-	"bytes"
 	"context"
 	"encoding/hex"
 	"errors"
@@ -126,151 +125,6 @@ func (m msgServer) EditSophon(goCtx context.Context, msg *types.MsgEditSophon) (
 	ctx.EventManager().EmitEvent(createSophonEvent(types.EventTypeUpdateSophon, sophonInfo))
 
 	return &types.MsgEditSophonResponse{}, nil
-}
-
-func (m msgServer) TransferOwnership(goCtx context.Context, msg *types.MsgTransferOwnership) (*types.MsgTransferOwnershipResponse, error) {
-	ctx := sdk.UnwrapSDKContext(goCtx)
-
-	pubKeyBytes, err := hex.DecodeString(msg.SophonPublicKey)
-	if err != nil {
-		return nil, errorsmod.Wrapf(err, "invalid hex in pubkey: %s", msg.SophonPublicKey)
-	}
-
-	sophonInfo, err := m.GetSophonInfo(ctx, pubKeyBytes)
-	if err != nil {
-		if errors.Is(err, collections.ErrNotFound) {
-			return nil, sdkerrors.ErrNotFound.Wrapf("sophon not found for %s", msg.SophonPublicKey)
-		}
-		return nil, err
-	}
-
-	if sophonInfo.OwnerAddress != msg.OwnerAddress {
-		return nil, sdkerrors.ErrorInvalidSigner.Wrapf("unauthorized owner; expected %s, got %s", sophonInfo.OwnerAddress, msg.OwnerAddress)
-	}
-
-	newOwnerAddress, err := sdk.AccAddressFromBech32(msg.NewOwnerAddress)
-	if err != nil {
-		return nil, sdkerrors.ErrInvalidAddress.Wrapf("invalid new owner address: %s", msg.NewOwnerAddress)
-	}
-
-	transferAddress, err := m.GetSophonTransfer(ctx, sophonInfo.Id)
-	if err != nil {
-		return nil, err
-	}
-	if transferAddress != nil {
-		return nil, sdkerrors.ErrInvalidRequest.Wrapf("there is already a pending ownership transfer for %s", msg.NewOwnerAddress)
-	}
-
-	err = m.SetSophonTransfer(ctx, sophonInfo.Id, newOwnerAddress)
-	if err != nil {
-		return nil, err
-	}
-
-	ctx.EventManager().EmitEvents(sdk.Events{
-		sdk.NewEvent(
-			types.EventTypeTransferOwnership,
-			sdk.NewAttribute(types.AttributeSophonPubKey, msg.SophonPublicKey),
-			sdk.NewAttribute(types.AttributeOwnerAddress, msg.OwnerAddress),
-			sdk.NewAttribute(types.AttributeNewOwnerAddress, msg.NewOwnerAddress),
-		),
-	})
-
-	return &types.MsgTransferOwnershipResponse{}, nil
-}
-
-func (m msgServer) AcceptOwnership(goCtx context.Context, msg *types.MsgAcceptOwnership) (*types.MsgAcceptOwnershipResponse, error) {
-	ctx := sdk.UnwrapSDKContext(goCtx)
-
-	pubKeyBytes, err := hex.DecodeString(msg.SophonPublicKey)
-	if err != nil {
-		return nil, errorsmod.Wrapf(err, "invalid hex in pubkey: %s", msg.SophonPublicKey)
-	}
-
-	sophonInfo, err := m.GetSophonInfo(ctx, pubKeyBytes)
-	if err != nil {
-		if errors.Is(err, collections.ErrNotFound) {
-			return nil, sdkerrors.ErrNotFound.Wrapf("sophon not found for %s", msg.SophonPublicKey)
-		}
-		return nil, err
-	}
-
-	newOwnerAddress, err := sdk.AccAddressFromBech32(msg.NewOwnerAddress)
-	if err != nil {
-		return nil, sdkerrors.ErrInvalidAddress.Wrapf("invalid new owner address: %s", msg.NewOwnerAddress)
-	}
-
-	transferAddress, err := m.GetSophonTransfer(ctx, sophonInfo.Id)
-	if err != nil {
-		return nil, err
-	}
-	if transferAddress == nil || !bytes.Equal(transferAddress, newOwnerAddress) {
-		return nil, sdkerrors.ErrNotFound.Wrapf("there is no pending ownership transfer for %s", msg.NewOwnerAddress)
-	}
-
-	err = m.DeleteSophonTransfer(ctx, sophonInfo.Id)
-	if err != nil {
-		return nil, err
-	}
-
-	sophonInfo.OwnerAddress = msg.NewOwnerAddress
-	err = m.SetSophonInfo(ctx, sophonInfo.PublicKey, sophonInfo)
-	if err != nil {
-		return nil, err
-	}
-
-	ctx.EventManager().EmitEvents(sdk.Events{
-		sdk.NewEvent(
-			types.EventTypeAcceptOwnership,
-			sdk.NewAttribute(types.AttributeSophonPubKey, msg.SophonPublicKey),
-			sdk.NewAttribute(types.AttributeOwnerAddress, msg.NewOwnerAddress),
-		),
-		createSophonEvent(types.EventTypeUpdateSophon, sophonInfo),
-	})
-
-	return &types.MsgAcceptOwnershipResponse{}, nil
-}
-
-func (m msgServer) CancelOwnershipTransfer(goCtx context.Context, msg *types.MsgCancelOwnershipTransfer) (*types.MsgCancelOwnershipTransferResponse, error) {
-	ctx := sdk.UnwrapSDKContext(goCtx)
-
-	pubKeyBytes, err := hex.DecodeString(msg.SophonPublicKey)
-	if err != nil {
-		return nil, errorsmod.Wrapf(err, "invalid hex in pubkey: %s", msg.SophonPublicKey)
-	}
-
-	sophonInfo, err := m.GetSophonInfo(ctx, pubKeyBytes)
-	if err != nil {
-		if errors.Is(err, collections.ErrNotFound) {
-			return nil, sdkerrors.ErrNotFound.Wrapf("sophon not found for %s", msg.SophonPublicKey)
-		}
-		return nil, err
-	}
-
-	if sophonInfo.OwnerAddress != msg.OwnerAddress {
-		return nil, sdkerrors.ErrorInvalidSigner.Wrapf("unauthorized owner; expected %s, got %s", sophonInfo.OwnerAddress, msg.OwnerAddress)
-	}
-
-	transferAddress, err := m.GetSophonTransfer(ctx, sophonInfo.Id)
-	if err != nil {
-		return nil, err
-	}
-	if transferAddress == nil {
-		return nil, sdkerrors.ErrNotFound.Wrapf("there is no pending ownership transfer for %s", msg.SophonPublicKey)
-	}
-
-	err = m.DeleteSophonTransfer(ctx, sophonInfo.Id)
-	if err != nil {
-		return nil, err
-	}
-
-	ctx.EventManager().EmitEvents(sdk.Events{
-		sdk.NewEvent(
-			types.EventTypeCancelOwnershipTransfer,
-			sdk.NewAttribute(types.AttributeSophonPubKey, msg.SophonPublicKey),
-		),
-	})
-
-	return &types.MsgCancelOwnershipTransferResponse{}, nil
 }
 
 func (m msgServer) AddUser(_ context.Context, _ *types.MsgAddUser) (*types.MsgAddUserResponse, error) {
